avances grales

+++ vista/IVistaPedido.java
@@ -9,7 +9,7 @@ public interface IVistaPedido { //La vista en la que se genera el pedido
 	 public void setControlador(Controlador controlador);
 	 public int getCantidadPasajeros();
 	 public String getTexto();
-	 public float getKm();
+	 public double getKm();
 	 public boolean getBaul();
 	 public boolean getMascota();
 }
